requirements.txt
Authlib==0.14.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==2.3.2
Flask-Cors==4.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pycparser==2.22
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.1

python-dotenv~=1.0.1
